Complete removal of the hero. Removal of text type. Container to the entirity of the site. Focus on about.
