--- src/components/TextType.vue
@@ -1,95 +1,0 @@
-/*
- * Created on Tue Mar 31 2020 by Johnathan Irvin
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-<template lang='pug'>
-  p.
-    {{written}}#[span(:class="{ hidden: !isVisible }") _]#[span(:class="{ hidden: isVisible }") {{'\u00A0'}}]
-</template>
-
-<script>
-export default {
-    props: {
-        text: {
-            type: String,
-            default: ''
-        },
-        hideCursor: {
-            type: Boolean,
-            default: false
-        },
-        typingSpeed: {
-            type: Number,
-            // During a light test I typed 485 chars in 1 min.
-            // This is around 8 characters per second --- around 95 WPM.
-            // My maximum typing speed is 140 WPM, but we'll use this
-            // as my typing speed is already fast as is.
-            // 1 Second / Character Per Minute == 1 key per X milliseconds.
-            default: 1000 / (485 / 60), // Typing speed in characters per millisecond. 
-        },
-        blinkSpeed: {
-            type: Number,
-            default: 1000 // Cursor Blink speed in milliseconds.
-        }, 
-        typeCharacters: {
-            type: Boolean,
-            default: true
-        }
-    },
-    data: function() {
-        return {
-            isVisible: false,
-            written: '',
-            finished: false
-        }
-    },
-    methods: {
-        ran: function(low, high) {
-             return Math.random() * low + high;
-        },
-        incrementWord: function() {
-            if (this.finished || !this.typeCharacters) {
-                if(!this.typeCharacters && !this.finished) {
-                    setTimeout(this.incrementWord, this.typingSpeed + this.ran(10, 20)); 
-                }
-                return;
-            } else {
-                setTimeout(this.incrementWord, this.typingSpeed + this.ran(10, 20)); 
-            }
-            let lenWritten = this.written.length;
-            this.written += this.text[lenWritten];
-            if (this.text == this.written) {
-                this.$emit('typed');
-                this.finished = true;
-            }
-        },
-        blink: function() {
-            this.isVisible = !this.isVisible && !this.hideCursor && this.typeCharacters;
-            setTimeout(this.blink, this.blinkSpeed); 
-        }
-    },
-    created: function() {
-        setTimeout(this.incrementWord, this.typingSpeed + this.ran(10, 20)); 
-        setTimeout(this.blink, this.blinkSpeed); 
-    }
-}
-</script>
-
-<style>
-.hidden {
-    display: none;
-}
-</style>
